+++ src/ast/passes/codegen_llvm.h
@@ -172,7 +172,9 @@ private:
                  FunctionType *func_type);
 
   [[nodiscard]] ScopedExprDeleter accept(Node *node);
-  [[nodiscard]] std::tuple<Value *, ScopedExprDeleter> getMapKey(Map &map);
+  [[nodiscard]] std::tuple<Value *, ScopedExprDeleter> getMapKey(
+      Map &map,
+      Expression *key_expr);
   AllocaInst *getMultiMapKey(Map &map, const std::vector<Value *> &extra_keys);
 
   void compareStructure(SizedType &our_type, llvm::Type *llvm_type);